requirements.txt
gitpython
openai
requests
schedule
